net: unify skb queue mapping API for RX and TX
For TX and RX, skb->queue_mapping has different interpretations.

On TX side, skb->queue_mapping is read and written directly,
the API's don't hide anything.

However, on RX side, skb->queue_mapping == 0 means the value
is not set, for all skb->queue_mapping > 0, it means the
index of the queue, with offset 1.

This should be okay as long as we don't use skb->queue_mapping
acorss RX->TX. But we do use this in netdev_pick_tx() for routers.

This would make skb->queue_mapping == 0 confusing when we choose
TX queue 0, netdev_pick_tx() will consider it not set and recompute
the hash.

This patch make the TX side align with RX side, which is more flexible,
this is safe except the following things need to fix:

1) We need to a new API to clear this field, rather than assigning
   0 to skb->queue_mapping directly

2) All existing skb_set_queue_mapping(skb, 0) need to review
   since it may mean to clear this field (RX side)

3) All existing !skb_get_queue_mapping(skb, 0) need to review
   since it may mean to test if it is set (bonding and team).

Signed-off-by: Cong Wang <xiyou.wangcong@gmail.com>