udp_tunnel: create struct sock instead of struct socket
We create udp tunnel sockets in kernel-space, there
is no reason to maintain the user-space part, that is struct socket,
struct sock should be enough for kernel to use.

Thanks to Tom's work on the new API, now we can save a lot
of API burden to convert to struct sock.

Cc: Tom Herbert <tom@herbertland.com>
Signed-off-by: Cong Wang <xiyou.wangcong@gmail.com>