bpf: introduce bpf timer
This patch introduces a bpf timer map and a syscall to create bpf timer
from user-space.

The reason why we have to use a map is because of the lifetime of a timer.
Without a map, we have to delete the timer before exiting the eBPF program,
this would significantly limit its use cases. With a map, the timer can
stay as long as the map itself and can be actually updated via map update
API's too (except creation), where the key is the timer ID and the value
is the timer expire time.

Timer creation is not easy to design either. In order to prevent users
creating a timer without adding it to a map, we have to enforce this in
the API which takes map as a parameter and adds the new timer into the
map in one shot.

And because timer is asynchronous, we can not just use its callback like
bpf_for_each_map_elem()'s. More importantly, we have to properly reference
count its struct bpf_prog too. It seems impossible to do this either in
the verifier or in JIT, so we have to make its callback code a separate
eBPF program and pass a program fd from user-space. This means we can not
create timers in eBPF code. Fortunately, timer callback can still live in
the same object file with the rest eBPF code and share data too.

Here is a quick (silly) demo of the timer callback code, to show how to
implement a timeout hashmap:

static __u64
check_expired_elem(struct bpf_map *map, __u32 *key, __u64 *val,
                  int *data)
{
  u64 expires = *val;

  if (expires < bpf_jiffies64()) {
    bpf_map_delete_elem(map, key);
    *data++;
  }
  return 0;
}

SEC("timer")
u32 timer_callback(void)
{
  int count = 0;

  bpf_for_each_map_elem(&map, check_expired_elem, &count, 0);
  if (count)
     return 0; // not re-arm this timer
  else
     return 10; // reschedule this timer after 10 jiffies
}

Cc: Alexei Starovoitov <ast@kernel.org>
Cc: Daniel Borkmann <daniel@iogearbox.net>
Cc: Andrii Nakryiko <andrii@kernel.org>
Cc: Martin KaFai Lau <kafai@fb.com>
Cc: Song Liu <songliubraving@fb.com>
Cc: Yonghong Song <yhs@fb.com>
Signed-off-by: Cong Wang <cong.wang@bytedance.com>