wireguard: preserve skb->mark on ingress side
wg_reset_packet() resets skb->mark twice: with skb_scrub_packet()
(xnet==true) and with memset() following it. But skb->mark does
not have to be cleared at least when staying in the same namespace,
other tunnels preserve it too.

In our use case, we would like to preserve this skb->mark to
distinguish which wireguard device the packets are routed from.

Cc: "Jason A. Donenfeld" <Jason@zx2c4.com>
Cc: wireguard@lists.zx2c4.com
Signed-off-by: Cong Wang <cong.wang@bytedance.com>